refactor: preparation of editor package separation
